split cache to files
